fix(npm): fix infinite loop by moving devtools to devdeps
