+++ addon/utils/datastore/prose-store.ts
@@ -3,7 +3,7 @@ import {
   RdfaParseConfig,
   RdfaParser,
 } from '@lblod/ember-rdfa-editor/utils/rdfa-parser/rdfa-parser';
-import { Node as PNode } from 'prosemirror-model';
+import { Node as PNode, ResolvedPos } from 'prosemirror-model';
 import { defaultPrefixes } from '@lblod/ember-rdfa-editor/config/rdfa';
 import { EditorState } from 'prosemirror-state';
 import SetUtils from '@lblod/ember-rdfa-editor/utils/set-utils';
@@ -11,28 +11,44 @@ import Datastore, {
   EditorStore,
 } from '@lblod/ember-rdfa-editor/utils/datastore/datastore';
 
-export interface ProseDatastore extends Datastore<PNode> {
+export type ResolvedPNode = {
+  node: PNode;
+  pos?: ResolvedPos;
+};
+
+export interface ProseDatastore extends Datastore<ResolvedPNode> {
   limitToRange(state: EditorState, start: number, end: number): ProseStore;
 }
 
-export class ProseStore extends EditorStore<PNode> implements ProseDatastore {
+export class ProseStore
+  extends EditorStore<ResolvedPNode>
+  implements ProseDatastore
+{
   limitToRange(state: EditorState, start: number, end: number): ProseStore {
-    const contextNodes = new Set();
+    const contextNodes: Set<PNode> = new Set();
     state.doc.nodesBetween(start, end, (node) => {
       contextNodes.add(node);
     });
+    console.log('CONTEXT NODES: ', contextNodes);
 
     return this.transformDataset((dataset) => {
       return dataset.filter((quad) => {
         const quadNodes = this._quadToNodes.get(quadHash(quad));
         if (quadNodes) {
           const { subjectNodes, predicateNodes, objectNodes } = quadNodes;
-          const hasSubjectNode = SetUtils.hasAny(contextNodes, ...subjectNodes);
+          const hasSubjectNode = SetUtils.hasAny(
+            contextNodes,
+            ...subjectNodes.map((resolvedNode) => resolvedNode.node)
+          );
           const hasPredicateNode = SetUtils.hasAny(
             contextNodes,
-            ...predicateNodes
+            ...predicateNodes.map((resolvedNode) => resolvedNode.node)
           );
-          const hasObjectNode = SetUtils.hasAny(contextNodes, ...objectNodes);
+          const hasObjectNode = SetUtils.hasAny(
+            contextNodes,
+            ...objectNodes.map((resolvedNode) => resolvedNode.node)
+          );
+
           return hasSubjectNode && hasPredicateNode && hasObjectNode;
         } else {
           return false;
@@ -42,7 +58,7 @@ export class ProseStore extends EditorStore<PNode> implements ProseDatastore {
   }
 }
 
-export function proseStoreFromParse(config: RdfaParseConfig<PNode>) {
+export function proseStoreFromParse(config: RdfaParseConfig<ResolvedPNode>) {
   const {
     dataset,
     subjectToNodesMapping,
