+++ package.json
@@ -71,7 +71,6 @@
     "iter-tools": "^7.4.0",
     "mdn-polyfills": "^5.16.0",
     "prosemirror-commands": "^1.3.1",
-    "prosemirror-dev-tools": "^3.1.0",
     "prosemirror-dropcursor": "^1.6.1",
     "prosemirror-gapcursor": "^1.3.1",
     "prosemirror-history": "^1.3.0",
@@ -162,6 +161,7 @@
     "loader.js": "^4.7.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.6.1",
+    "prosemirror-dev-tools": "^3.1.0",
     "qunit": "^2.18.1",
     "qunit-dom": "^2.0.0",
     "release-it": "^15.4.1",
